chore: rename fetch fixture file
The exported function is makeFetchRequest, makeRangeRequest was
refactored away previously so update the file name to reflect this.
