+++ test-e2e/dag-pb.test.ts
@@ -2,7 +2,7 @@ import { stop } from '@libp2p/interface'
 import { createKuboRPCClient } from 'kubo-rpc-client'
 import { test, expect } from './fixtures/config-test-fixtures.ts'
 import { loadWithServiceWorker } from './fixtures/load-with-service-worker.ts'
-import { makeFetchRequest } from './fixtures/make-range-request.ts'
+import { makeFetchRequest } from './fixtures/make-fetch-request.ts'
 import { loadBypassingMediaViewer } from './fixtures/media-viewer.ts'
 import type { KuboRPCClient } from 'kubo-rpc-client'
